fix(site): :lock: js/xss-through-dom vulnerability
Extracting text from a DOM node and interpreting it as HTML can lead to a cross-site scripting vulnerability.  A webpage with this vulnerability reads text from the DOM, and afterwards adds the text as HTML to the DOM. Using text from the DOM as HTML effectively unescapes the text, and thereby invalidates any escaping done on the text. If an attacker is able to control the safe sanitized text, then this vulnerability can be exploited to perform a cross-site scripting attack.